tehtävä 21*

+++ osa4/controllers/blogs.ts
@@ -1,9 +1,15 @@
 import { NextFunction, Request, Response } from "express";
 import { BlogT } from "../models/blog";
 import { UserT } from "../models/user";
+import jwt, { JwtPayload } from "jsonwebtoken";
 const BlogModel = require("../models/blog");
 const UserModel = require("../models/user");
 const blogRouter = require("express").Router();
+
+interface TokenRequest extends Request {
+  token?: string;
+  user: UserT;
+}
 
 blogRouter.get("/", async (request: Request, response: Response) => {
   const blogs: Array<BlogT> = await BlogModel.find({}).populate("user", {
@@ -28,10 +34,14 @@ blogRouter.get(
 
 blogRouter.post(
   "/",
-  async (request: Request, response: Response, next: NextFunction) => {
+  async (request: TokenRequest, response: Response, next: NextFunction) => {
     const body: BlogT = request.body;
+    const decodedToken = jwt.verify(request.token!, process.env.SECRET!) as any;
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
 
-    const user: UserT = await UserModel.findById(body.userId);
+    const user: UserT = await UserModel.findById(decodedToken.id);
 
     const blog = new BlogModel({
       title: body.title,
@@ -50,9 +60,19 @@ blogRouter.post(
 
 blogRouter.delete(
   "/:id",
-  async (request: Request, response: Response, next: NextFunction) => {
-    await BlogModel.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+  async (request: TokenRequest, response: Response, next: NextFunction) => {
+    const decodedToken = jwt.verify(request.token!, process.env.SECRET!) as any;
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+    const user: UserT = await UserModel.findById(decodedToken.id);
+    const blog: BlogT | null = await BlogModel.findById(request.params.id);
+    if (blog!.user!.toString() === user._id!.toString()) {
+      await BlogModel.findByIdAndDelete(request.params.id);
+      response.status(204).end();
+    } else {
+      return response.status(401).json({ error: "unauthorized" });
+    }
   }
 );
